requirements.txt
opencv-python
ultralytics
numpy
python-dotenv
langchain-groq
langgraph
pytest
